+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The git repositories are hosted at the following sites:
 
 ### Added
 
+* Added API and optional properties to basic lighting output object example
+  for power and feedback value. (#1185)
 * Added properties to the Channel object write member value coercion
   minimal properties supported. (#1176)
 * Added Send_x_Address() API to ReadPropertyMultiple, WritePropertyMultiple,